Add middleware
This adds request id, session, and csrf middleware along with tests.
Session middleware is required in order to use the csrf solution and
currently uses the gorilla session implementation. We'll probably stick
with this for now but it would be nice to completely replace that as
well. The gorilla project seems to be semi-unmaintained and unreliable
which is the reason for releasing this csrf middleware to begin with.
They broke compatibility with the introduction of the origin list which
is only necessary for CORS.